Implement mobile-friendly design
By manipulating the flex-grid offsets slightly we can get a mobile-friendly
design without changing the structure much at all.

Mostly this patch removes some empty divs that were used for spacing. Empty divs
really shouldn't be used for layout anyway.

On mobile, the end result is a usable interface where all the elements are
accessible. On desktop, the end result is that the app is slightly more
left-aligned, which I think is a fair trade-off. Everything else seems correctly
aligned as before.

The only thing I didn't get to test was the Forum, which I don't know how to
use.